src/requirements.txt
azure-functions
azure-functions-durable
azure-data-tables